refactor: 서비스 레이어 정의
