+++ src/main/java/chess/dao/DatabaseMemberDao.java
@@ -1,0 +1,67 @@
+package chess.dao;
+
+import chess.domain.Member;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
+
+public class DatabaseMemberDao implements MemberDao {
+
+    @Override
+    public Long save(Member member) {
+        final Connection connection = MysqlConnector.getConnection();
+        final String sql = "insert into Member (name) values (?)";
+        try {
+            final PreparedStatement statement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
+            statement.setString(1, member.getName());
+            statement.executeUpdate();
+            ResultSet resultSet = statement.getGeneratedKeys();
+            if (resultSet.next()) {
+                return resultSet.getLong(1);
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return 0L;
+    }
+
+    @Override
+    public Optional<Member> findById(Long id) {
+        final Connection connection = MysqlConnector.getConnection();
+        final String sql = "select id, name from Member where id = ?";
+        try {
+            final PreparedStatement statement = connection.prepareStatement(sql);
+            statement.setLong(1, id);
+            ResultSet resultSet = statement.executeQuery();
+            if (!resultSet.next()) {
+                return Optional.empty();
+            }
+            return Optional.of(new Member(resultSet.getLong("id"), resultSet.getString("name")));
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return Optional.empty();
+    }
+
+    @Override
+    public List<Member> findAll() {
+        final Connection connection = MysqlConnector.getConnection();
+        final String sql = "select id, name from Member";
+        List<Member> members = new ArrayList<>();
+        try {
+            final PreparedStatement statement = connection.prepareStatement(sql);
+            ResultSet resultSet = statement.executeQuery();
+            while (resultSet.next()) {
+                members.add(new Member(resultSet.getLong("id"), resultSet.getString("name")));
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return members;
+    }
+}
