+++ src/main/java/chess/dao/MemberDao.java
@@ -1,57 +1,13 @@
 package chess.dao;
 
 import chess.domain.Member;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
-public class MemberDao {
+public interface MemberDao {
+    Long save(Member member);
 
-    public void save(Member member) {
-        final Connection connection = MysqlConnector.getConnection();
-        final String sql = "insert into Member (name) values (?)";
-        try {
-            final PreparedStatement statement = connection.prepareStatement(sql);
-            statement.setString(1, member.getName());
-            statement.executeUpdate();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-    }
+    Optional<Member> findById(Long id);
 
-    public Member findById(final Long id) {
-        final Connection connection = MysqlConnector.getConnection();
-        final String sql = "select id, name from Member where id = ?";
-        try {
-            final PreparedStatement statement = connection.prepareStatement(sql);
-            statement.setLong(1, id);
-            ResultSet resultSet = statement.executeQuery();
-            if (!resultSet.next()) {
-                return null;
-            }
-            return new Member(resultSet.getLong("id"), resultSet.getString("name"));
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        return null;
-    }
-
-    public List<Member> findAll() {
-        final Connection connection = MysqlConnector.getConnection();
-        final String sql = "select id, name from Member";
-        List<Member> members = new ArrayList<>();
-        try {
-            final PreparedStatement statement = connection.prepareStatement(sql);
-            ResultSet resultSet = statement.executeQuery();
-            while (resultSet.next()) {
-                members.add(new Member(resultSet.getLong("id"), resultSet.getString("name")));
-            }
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        return members;
-    }
+    List<Member> findAll();
 }
