+++ src/main/java/chess/dao/MemoryMemberDao.java
@@ -1,4 +1,4 @@
-package chess.repository;
+package chess.dao;
 
 import chess.domain.Member;
 import java.util.ArrayList;
@@ -7,19 +7,20 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 
-public class MemoryMemberRepository implements MemberRepository {
+public class MemoryMemberDao implements MemberDao {
     private static Map<Long, Member> store = new ConcurrentHashMap<>();
     private static int nextId = 1;
 
     @Override
-    public void save(Member member) {
+    public Long save(Member member) {
         member = new Member((long) nextId++, member.getName());
         store.put(member.getId(), member);
+        return member.getId();
     }
 
     @Override
     public Optional<Member> findById(Long id) {
-        return Optional.of(store.get(id));
+        return Optional.ofNullable(store.get(id));
     }
 
     @Override
