+++ src/main/java/chess/domain/ChessGame.java
@@ -19,6 +19,11 @@ public class ChessGame {
         this.participant = participant;
     }
 
+    public ChessGame(Long id, GameState state) {
+        this.id = id;
+        this.state = state;
+    }
+
     public ChessGame(Long id, GameState state, Participant participant) {
         this.id = id;
         this.state = state;
@@ -101,5 +106,9 @@ public class ChessGame {
 
     public Long getBlackId() {
         return participant.getBlackId();
+    }
+
+    public void participate(Member white, Member black) {
+        this.participant = new Participant(white, black);
     }
 }
