+++ src/main/resources/templates/play.html
@@ -26,11 +26,19 @@
         }
 
         table.chess-board > tbody > tr:nth-child(2n) > td:nth-child(2n) {
-            background-color: #aaa;
+            background-color: gray;
+        }
+
+        table.chess-board > tbody > tr:nth-child(2n) > td:nth-child(2n-1) {
+            background-color: wheat;
         }
 
         table.chess-board > tbody > tr:nth-child(2n-1) > td:nth-child(2n-1) {
-            background-color: #aaa;
+            background-color: gray;
+        }
+
+        table.chess-board > tbody > tr:nth-child(2n-1) > td:nth-child(2n) {
+            background-color: wheat;
         }
 
         img.chess-piece-img {
